+++ packages/snap/src/constants/index.ts
@@ -3,6 +3,7 @@ import { BigNumber } from 'bignumber.js';
 
 export const ZERO = BigNumber(0);
 export const ACCOUNT_ACTIVATION_FEE_TRX = BigNumber(1);
+export const MEMO_FEE_TRX = BigNumber(1);
 export const SUN_IN_TRX = 1_000_000;
 /**
  * 100 TRX